--- a/src/init.c
+++ b/src/init.c
@@ -6,7 +6,7 @@
 /* .Call calls */
 extern SEXP slide_common_impl(SEXP, SEXP, SEXP, SEXP, SEXP);
 extern SEXP slide_between_common_impl(SEXP, SEXP, SEXP, SEXP, SEXP, SEXP, SEXP, SEXP, SEXP, SEXP, SEXP);
-extern SEXP slide_index_common_impl(SEXP, SEXP, SEXP, SEXP, SEXP, SEXP, SEXP, SEXP, SEXP);
+extern SEXP slide_index_common_impl(SEXP, SEXP, SEXP, SEXP, SEXP, SEXP, SEXP, SEXP, SEXP, SEXP, SEXP, SEXP);
 
 // Defined below
 SEXP slide_init(SEXP);
@@ -14,7 +14,7 @@ SEXP slide_init(SEXP);
 static const R_CallMethodDef CallEntries[] = {
   {"slide_common_impl",         (DL_FUNC) &slide_common_impl, 5},
   {"slide_between_common_impl", (DL_FUNC) &slide_between_common_impl, 11},
-  {"slide_index_common_impl",   (DL_FUNC) &slide_index_common_impl, 9},
+  {"slide_index_common_impl",   (DL_FUNC) &slide_index_common_impl, 12},
   {"slide_init",                (DL_FUNC) &slide_init, 1},
   {NULL, NULL, 0}
 };