… src/slide.c