Rewrote the guts of `slide_index()` and `slide_between()`. The main improvements are that:

- Many of the structs have been removed in favor of simpler implementations
- The biggest benefit is that we no longer call `vec_slice_impl()` when we do any window updating. This really hurt performance with all of the allocations.

Before

``` r
library(slide)

x <- 1:1e5 + 0L
y <- x - 10L

bench::mark(
  slide(x, identity, .before = 10L),
  slide_index(x, x, identity, .before = 10L),
  slide_between(x, x, identity, .starts = y, .stops = x),
  iterations = 100
)
#> Warning: Some expressions had a GC in every iteration; so filtering is
#> disabled.
#> # A tibble: 3 x 6
#>   expression                                                 min  median
#>   <bch:expr>                                             <bch:t> <bch:t>
#> 1 slide(x, identity, .before = 10L)                       42.6ms  50.7ms
#> 2 slide_index(x, x, identity, .before = 10L)             103.1ms 127.9ms
#> 3 slide_between(x, x, identity, .starts = y, .stops = x) 100.8ms 127.2ms
#> # … with 3 more variables: `itr/sec` <dbl>, mem_alloc <bch:byt>,
#> #   `gc/sec` <dbl>
```

<sup>Created on 2019-11-19 by the [reprex package](https://reprex.tidyverse.org) (v0.3.0.9000)</sup>

After

```r
library(slide)

x <- 1:1e5 + 0L
y <- x - 10L

bench::mark(
  slide(x, identity, .before = 10L),
  slide_index(x, x, identity, .before = 10L),
  slide_between(x, x, identity, .starts = y, .stops = x),
  iterations = 100
)
#> Warning: Some expressions had a GC in every iteration; so filtering is
#> disabled.
#> # A tibble: 3 x 6
#>   expression                                                min median
#>   <bch:expr>                                             <bch:> <bch:>
#> 1 slide(x, identity, .before = 10L)                      39.9ms 47.5ms
#> 2 slide_index(x, x, identity, .before = 10L)             63.3ms 74.6ms
#> 3 slide_between(x, x, identity, .starts = y, .stops = x) 63.1ms 74.5ms
#> # … with 3 more variables: `itr/sec` <dbl>, mem_alloc <bch:byt>,
#> #   `gc/sec` <dbl>
```